Moved edit code into its own class
