+++ lib/frameit.rb
@@ -6,6 +6,7 @@ require 'frameit/runner'
 require 'frameit/screenshot'
 require 'frameit/config_parser'
 require 'frameit/offsets'
+require 'frameit/editor'
 require 'frameit/template_finder'
 require 'frameit/strings_parser'
 require 'frameit/dependency_checker'
